Use a sans-serif font for the over-unit text.
With the stroke and other visuals going on around it, the serifs just
make things busier and harder to read.

This progress on #2, but we're not there yet; in spite of the outline
the text still sometimes clashes with the background.  Perhaps I should
just put a background on the text, so we control the bg color.